+++ cpu_func.py
@@ -4,12 +4,12 @@ def modeAbsoluteFunc(cpu):
     cpu.address = cpu.Read16(cpu.PC + uint16(1))
 
 def modeAbsoluteXFunc(cpu):
-    cpu.address = cpu.Read16(cpu.PC + uint16(1)) + cpu.X
-    cpu.pageCrossed = cpu.pagesDiffer(cpu.address - cpu.X, cpu.address)
+    cpu.address = cpu.Read16(cpu.PC + uint16(1)) + uint16(cpu.X)
+    cpu.pageCrossed = cpu.pagesDiffer(cpu.address - uint16(cpu.X), cpu.address)
 
 def modeAbsoluteYFunc(cpu):
-    cpu.address = cpu.Read16(cpu.PC + uint16(1)) + cpu.Y   
-    cpu.pageCrossed = cpu.pagesDiffer(cpu.address - cpu.Y, cpu.address)
+    cpu.address = cpu.Read16(cpu.PC + uint16(1)) + uint16(cpu.Y)
+    cpu.pageCrossed = cpu.pagesDiffer(cpu.address - uint16(cpu.Y), cpu.address)
 
 def modeAccumulatorFunc(cpu):
     cpu.address = uint16(0)
@@ -21,7 +21,7 @@ def modeImpliedFunc(cpu):
     cpu.address = uint16(0)
 
 def modeIndexedIndirectFunc(cpu):
-    cpu.address = cpu.read16bug(uint16(cpu.Read(cpu.PC + uint16(1))) + cpu.X)
+    cpu.address = cpu.read16bug(uint16(cpu.Read(cpu.PC + uint16(1))) + uint16(cpu.X))
 
 def modeIndirectFunc(cpu):
     cpu.address = cpu.read16bug(cpu.Read16(cpu.PC + uint16(1)))
@@ -41,7 +41,7 @@ def modeZeroPageFunc(cpu):
     cpu.address = uint16(cpu.Read(cpu.PC + uint16(1)))
 
 def modeZeroPageX(cpu):
-    cpu.address = uint16(cpu.Read(cpu.PC + uint16(1)) + cpu.X)
+    cpu.address = uint16(cpu.Read(cpu.PC + uint16(1)) + uint16(cpu.X))
 
 def modeZeroPageY(cpu):
-    cpu.address = uint16(cpu.Read(cpu.PC + uint16(1)) + cpu.Y)
+    cpu.address = uint16(cpu.Read(cpu.PC + uint16(1)) + uint16(cpu.Y))
